[REFACTOR] Players 예외처리 (5단계)
