+++ src/main/java/domain/LadderGame.java
@@ -1,18 +1,6 @@
 package domain;
 
-import java.util.ArrayList;
-import java.util.List;
-
 public class LadderGame {
-
-    public List<Player> createPlayer(List<String> playerNames) {
-        List<Player> players = new ArrayList<>();
-        for (int i = 0; i < playerNames.size(); i++) {
-            players.add(new Player(new Name(playerNames.get(i)), new Position(i)));
-        }
-
-        return players;
-    }
 
     public void movePlayer(Line line, Player player) {
         if (line.canMoveRight(player.getPosition())) {
